+++ src/routes/(auth)/login/+page.server.ts
@@ -1,3 +1,4 @@
+import { loginSchema } from '$lib/schemas/loginSchema';
 import { createSession, generateSessionToken, setSessionTokenCookie } from '$lib/server/auth';
 import { providers } from '$lib/server/auth/oauth/OAuthClient';
 import db from '$lib/server/db';
@@ -5,32 +6,27 @@ import users from '$lib/server/db/schema/user';
 import { verifyPassword } from '$lib/server/utils';
 import { fail, redirect } from '@sveltejs/kit';
 import { eq } from 'drizzle-orm';
+import { StatusCodes } from 'http-status-codes';
 import { message, superValidate } from 'sveltekit-superforms';
 import { zod } from 'sveltekit-superforms/adapters';
-import { z } from 'zod';
 import type { Actions, PageServerLoad } from './$types';
-
-const schema = z.object({
-	email: z.string().email().max(128),
-	password: z.string().max(128)
-});
 
 export const load = (async (event) => {
 	if (event.locals.user) {
-		return redirect(302, '/profile');
+		return redirect(StatusCodes.TEMPORARY_REDIRECT, '/profile');
 	}
 
-	const form = await superValidate(zod(schema));
+	const form = await superValidate(zod(loginSchema));
 
 	return { form, providers };
 }) satisfies PageServerLoad;
 
 export const actions = {
 	login: async (event) => {
-		const form = await superValidate(event.request, zod(schema));
+		const form = await superValidate(event.request, zod(loginSchema));
 
 		if (!form.valid) {
-			return fail(400, { form });
+			return fail(StatusCodes.BAD_REQUEST, { form });
 		}
 
 		const { email, password } = form.data;
@@ -40,13 +36,13 @@ export const actions = {
 		const validPassword = await verifyPassword(existingUser?.password ?? '', password);
 
 		if (!validPassword || !existingUser) {
-			return message(form, { status: 'error', text: 'Incorrect username or password' });
+			return message(form, { status: 'error', text: 'Incorrect email or password' });
 		}
 
 		const sessionToken = generateSessionToken();
 		const session = await createSession(sessionToken, existingUser.id);
 		setSessionTokenCookie(event, sessionToken, session.expiresAt);
 
-		return redirect(302, '/profile');
+		return redirect(StatusCodes.TEMPORARY_REDIRECT, '/profile');
 	}
 } satisfies Actions;