+++ src/routes/+page.server.ts
@@ -1,9 +1,10 @@
 import { redirect } from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
+import { StatusCodes } from 'http-status-codes';
 
 export const load = (async ({ locals }) => {
 	if (locals.user) {
-		throw redirect(302, '/submissions');
+		throw redirect(StatusCodes.TEMPORARY_REDIRECT, '/submissions');
 	}
 	return {};
 }) satisfies PageServerLoad;