+++ src/routes/admin/+page.ts
@@ -1,6 +1,7 @@
 import { redirect } from '@sveltejs/kit';
 import type { PageLoad } from './$types';
+import { StatusCodes } from 'http-status-codes';
 
 export const load = (async () => {
-	return redirect(302, '/admin/events');
+	return redirect(StatusCodes.TEMPORARY_REDIRECT, '/admin/events');
 }) satisfies PageLoad;